feat: :poop: Improve logic for the get valid and invalid cards, play trump & non-trump cards, and the play cards method, but using the proper logic by getting the index of the ranks list, to determine the card order
Add a few more print statements for debugging purposes. Redistribution of cards still doesn't work as expected. Also, enhance the game's decision-making process for the AI regarding playing cards, focusing on deciding whether to throw a trump or the best non-trump card, as well as refining the logic for handling table cards.